Update readme to provide easy fix for a common problem
Quick and easy fix for using this plugin together with dark mode color schemes.
People who are not technical enough might not be able to figure this out on their own will benefit.
This is my first ever open source commit. I hope it will somehow get through.
Thanks for this amazing plugin by the way!